requirements.txt
Django==4.2.2
ipython==8.14.0
django-extensions==3.2.3
psycopg2-binary==2.9.6
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2


